# Use a small traces table for querying.