--- a/internal/status/templates/traces.html
+++ b/internal/status/templates/traces.html
@@ -48,7 +48,7 @@
             <tr>
               <th scope="col">Trace URL</th>
               <th scope="col">Start Time</th>
-              <th scope="col">Duration</th>
+              <th scope="col">Latency</th>
             </tr>
           </thead>
           <tbody>