--- a/runtime/perfetto/db.go
+++ b/runtime/perfetto/db.go
@@ -157,7 +157,7 @@ func (d *DB) Store(ctx context.Context, app, version string, spans *protos.Trace
 	if err != nil {
 		return err
 	}
-	defer tx.Rollback() //nolint:errcheck supplanted by errs below
+	defer tx.Rollback() //nolint:errcheck // supplanted by errs below
 	var errs []error
 	for _, span := range spans.Span {
 		if isRootSpan(span) {